+++ INTERX/config/type.go
@@ -41,6 +41,7 @@ type CacheConfig struct {
 
 // InterxConfig is a struct to be used for interx configuration
 type InterxConfig struct {
+	ServeHTTPS bool           `json:"serve_https"`
 	GRPC       string         `json:"grpc"`
 	RPC        string         `json:"rpc"`
 	PORT       string         `json:"port"`
@@ -55,6 +56,7 @@ type InterxConfig struct {
 
 // InterxConfigFromFile is a struct to be used for interx configuration file
 type InterxConfigFromFile struct {
+	ServeHTTPS   bool   `json:"serve_https"`
 	GRPC         string `json:"grpc"`
 	RPC          string `json:"rpc"`
 	PORT         string `json:"port"`