+++ INTERX/proto-gen/kira/gov/pagination.pb.go
@@ -40,18 +40,18 @@ type PageRequest struct {
 	// key is a value returned in PageResponse.next_key to begin
 	// querying the next page most efficiently. Only one of offset or key
 	// should be set.
-	Key []byte `protobuf:"bytes,1,opt,name=key,proto3" json:"key,omitempty"`
+	Key []byte `protobuf:"bytes,1,opt,name=key,proto3" json:"key"`
 	// offset is a numeric offset that can be used when key is unavailable.
 	// It is less efficient than using key. Only one of offset or key should
 	// be set.
-	Offset uint64 `protobuf:"varint,2,opt,name=offset,proto3" json:"offset,omitempty"`
+	Offset uint64 `protobuf:"varint,2,opt,name=offset,proto3" json:"offset"`
 	// limit is the total number of results to be returned in the result page.
 	// If left empty it will default to a value to be set by each app.
-	Limit uint64 `protobuf:"varint,3,opt,name=limit,proto3" json:"limit,omitempty"`
+	Limit uint64 `protobuf:"varint,3,opt,name=limit,proto3" json:"limit"`
 	// count_total is set to true  to indicate that the result set should include
 	// a count of the total number of items available for pagination in UIs. count_total
 	// is only respected when offset is used. It is ignored when key is set.
-	CountTotal bool `protobuf:"varint,4,opt,name=count_total,json=countTotal,proto3" json:"count_total,omitempty"`
+	CountTotal bool `protobuf:"varint,4,opt,name=count_total,json=countTotal,proto3" json:"count_total"`
 }
 
 func (x *PageRequest) Reset() {
@@ -128,10 +128,10 @@ type PageResponse struct {
 
 	// next_key is the key to be passed to PageRequest.key to
 	// query the next page most efficiently
-	NextKey []byte `protobuf:"bytes,1,opt,name=next_key,json=nextKey,proto3" json:"next_key,omitempty"`
+	NextKey []byte `protobuf:"bytes,1,opt,name=next_key,json=nextKey,proto3" json:"next_key"`
 	// total is total number of results available if PageRequest.count_total
 	// was set, its value is undefined otherwise
-	Total uint64 `protobuf:"varint,2,opt,name=total,proto3" json:"total,omitempty"`
+	Total uint64 `protobuf:"varint,2,opt,name=total,proto3" json:"total"`
 }
 
 func (x *PageResponse) Reset() {