+++ x/gov/keeper/grpc_query.go
@@ -109,7 +109,6 @@ func (q Querier) GetPoorNetworkMessages(ctx context.Context, request *types.Poor
 
 // Proposal returns a proposal by id
 func (q Querier) Proposal(ctx context.Context, request *types.QueryProposalRequest) (*types.QueryProposalResponse, error) {
-	fmt.Println("Proposal request", request)
 	sdkContext := sdk.UnwrapSDKContext(ctx)
 	proposal, found := q.keeper.GetProposal(sdkContext, request.ProposalId)
 	if found == false {
@@ -118,7 +117,7 @@ func (q Querier) Proposal(ctx context.Context, request *types.QueryProposalReque
 	votes := q.keeper.GetProposalVotes(sdkContext, request.ProposalId)
 	return &types.QueryProposalResponse{
 		Proposal: proposal,
-		Votes: votes,
+		Votes:    votes,
 	}, nil
 }